requirements.txt
ollama
langchain-community
langchain-ollama
pymupdf
chromadb
streamlit
